requirements.txt
openai>=1.50.0
pvporcupine>=3.0.0
sounddevice>=0.5.0
webrtcvad>=2.0.10
pydantic>=2.9.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
numpy>=1.26.0
google-api-python-client>=2.154.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
aiohttp>=3.10.2
websockets>=12.0
ruff>=0.6.4
pytest>=8.3.0
